+++ lib/widgets/modal_input_url.dart
@@ -7,11 +7,13 @@ class ModalInputUrl extends StatelessWidget {
     required this.toolbar,
     required this.leftText,
     required this.selection,
+    this.onActionCompleted,
   }) : super(key: key);
 
   final Toolbar toolbar;
   final String leftText;
   final TextSelection selection;
+  final VoidCallback? onActionCompleted;
 
   @override
   Widget build(BuildContext context) {
@@ -20,75 +22,64 @@ class ModalInputUrl extends StatelessWidget {
         bottom: MediaQuery.of(context).viewInsets.bottom,
       ),
       padding: const EdgeInsets.all(30),
-      width: double.infinity,
+      width: double.maxFinite,
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         mainAxisSize: MainAxisSize.min,
         children: [
-          const Text(
-            "Please provide a URL here.",
-            style: TextStyle(
-              fontSize: 18,
-              fontWeight: FontWeight.bold,
+          const Padding(
+            padding: EdgeInsets.all(8.0),
+            child: Text(
+              "Please provide a URL here.",
+              style: TextStyle(
+                fontSize: 18,
+                fontWeight: FontWeight.bold,
+              ),
             ),
           ),
-          Container(
-            margin: const EdgeInsets.only(top: 10, bottom: 15),
-            padding: const EdgeInsets.symmetric(
-              vertical: 15,
-              horizontal: 15,
-            ),
-            decoration: BoxDecoration(
-              border: Border.all(color: Colors.grey),
-              borderRadius: BorderRadius.circular(10),
-            ),
-            child: TextField(
-              autocorrect: false,
-              autofocus: true,
-              cursorRadius: const Radius.circular(20),
-              decoration: const InputDecoration.collapsed(
-                hintText: "Input your url.",
+          TextField(
+            autocorrect: false,
+            autofocus: true,
+            cursorRadius: const Radius.circular(16),
+            decoration: const InputDecoration(
+              hintText: "Input your url.",
+              helperText: "example: https://example.com",
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.all(Radius.circular(16.0)),
               ),
-              style: const TextStyle(
-                fontSize: 16,
-              ),
-              enableInteractiveSelection: true,
-              onSubmitted: (String value) {
-                Navigator.pop(context);
+            ),
+            style: const TextStyle(fontSize: 16),
+            enableInteractiveSelection: true,
+            onSubmitted: (String value) {
+              Navigator.pop(context);
 
-                /// check if the user entered an empty input
-                if (value.isEmpty) {
-                  ScaffoldMessenger.of(context).showSnackBar(
-                    SnackBar(
-                      content: const Text(
-                        "Please input url",
-                        style: TextStyle(
-                          color: Colors.white,
-                        ),
+              /// check if the user entered an empty input
+              if (value.isEmpty) {
+                ScaffoldMessenger.of(context).showSnackBar(
+                  SnackBar(
+                    content: const Text(
+                      "Please input url",
+                      style: TextStyle(
+                        color: Colors.white,
                       ),
-                      backgroundColor: Colors.red.withOpacity(0.8),
-                      duration: const Duration(milliseconds: 700),
                     ),
-                  );
-                } else {
-                  if (!value.contains(RegExp(r'https?:\/\/(www.)?([^\s]+)'))) {
-                    value = "http://$value";
-                  }
-                  toolbar.action(
-                    "$leftText$value)",
-                    "",
-                    textSelection: selection,
-                  );
+                    backgroundColor: Colors.red.withOpacity(0.8),
+                    duration: const Duration(milliseconds: 700),
+                  ),
+                );
+              } else {
+                if (!value.contains(RegExp(r'https?:\/\/(www.)?([^\s]+)'))) {
+                  value = "http://$value";
                 }
-              },
-            ),
-          ),
-          const Text(
-            "example: https://example.com",
-            style: TextStyle(
-              fontSize: 12,
-              color: Colors.grey,
-            ),
+                toolbar.action(
+                  "$leftText$value)",
+                  "",
+                  textSelection: selection,
+                );
+              }
+
+              onActionCompleted?.call();
+            },
           ),
         ],
       ),